DD-31 📁 [File] 공통 컴포넌트 정리 페이지 /guide 경로로 이동
