+++ src/routes/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import MainPage from '@/pages/MainPage';
+import MainPage from '@/pages/GuidePage';
 import LoginPage from '@/pages/LoginPage';
 import GroupPage from '@/pages/GroupPage';
 import ProfilePage from '@/pages/ProfilePage';
@@ -11,11 +11,16 @@ import SignUpEmailPage from '@/pages/signup/SignUpEmailPage';
 import SignUpProfilePage from '@/pages/signup/SignUpProfilePage';
 import SignupLayout from '@/layouts/SignupLayout';
 import SignUpCompletePage from '@/pages/signup/SignUpCompletePage';
+import GuidePage from '@/pages/GuidePage';
 
 const router = createBrowserRouter([
   {
     path: '/',
     element: <MainPage />,
+  },
+  {
+    path: 'guide',
+    element: <GuidePage />,
   },
   {
     path: 'login',
